Project structure reorganized
