+++ src/JavaScriptObfuscator.ts
@@ -14,11 +14,11 @@ import { IObfuscationEventEmitter } from './interfaces/event-emitters/IObfuscati
 import { IObfuscationResult } from './interfaces/IObfuscationResult';
 import { IOptions } from './interfaces/options/IOptions';
 import { IRandomGenerator } from './interfaces/utils/IRandomGenerator';
-import { ISourceMapCorrector } from './interfaces/ISourceMapCorrector';
+import { ISourceMapCorrector } from './interfaces/source-map/ISourceMapCorrector';
 import { IStackTraceAnalyzer } from './interfaces/analyzers/stack-trace-analyzer/IStackTraceAnalyzer';
 import { IStackTraceData } from './interfaces/analyzers/stack-trace-analyzer/IStackTraceData';
 import { IStorage } from './interfaces/storages/IStorage';
-import { ITransformersRunner } from './interfaces/ITransformersRunner';
+import { ITransformersRunner } from './interfaces/node-transformers/ITransformersRunner';
 
 import { LoggingMessage } from './enums/logger/LoggingMessage';
 import { NodeTransformer } from './enums/container/node-transformers/NodeTransformer';
