+++ src/enums/node/NodeType.ts
