+++ src/interfaces/source-map/ISourceMapCorrector.d.ts
@@ -1,4 +1,4 @@
-import { IObfuscationResult } from './IObfuscationResult';
+import { IObfuscationResult } from '../IObfuscationResult';
 
 export interface ISourceMapCorrector {
     /**
